pyproject.toml
[project]
name = "kleinanzeigen-agent"
version = "0.1.0"
description = "Telegram bot that drafts Kleinanzeigen listings from photos via Claude Vision"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[job-queue]>=21.6",
    "strands-agents>=0.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
kleinanzeigen-agent = "main:main"
